+++ testsuite/tests/th/T10828.hs
@@ -33,16 +33,28 @@ $( return
            [ PlainTV (mkName "a") ]
            (Just StarT)
            [ GadtC [(mkName "MkT")]
-                   [ (NotStrict, VarT (mkName "a"))
-                   , (NotStrict, VarT (mkName "a"))]
+                   [ ( Bang NoSourceUnpackedness NoSourceStrictness
+                     , VarT (mkName "a")
+                     )
+                   , ( Bang NoSourceUnpackedness NoSourceStrictness
+                     , VarT (mkName "a")
+                     )
+                   ]
                    ( mkName "T" )
                    [ VarT (mkName "a") ]
            , ForallC [PlainTV (mkName "a"), PlainTV (mkName "b")]
                      [AppT (AppT EqualityT (VarT $ mkName "a"  ) )
                                            (ConT $ mkName "Int") ] $
              RecGadtC [(mkName "MkC")]
-                  [ (mkName "foo", NotStrict, VarT (mkName "a"))
-                  , (mkName "bar", NotStrict, VarT (mkName "b"))]
+                  [ ( mkName "foo"
+                    , Bang NoSourceUnpackedness NoSourceStrictness
+                    , VarT (mkName "a")
+                    )
+                  , ( mkName "bar"
+                    , Bang NoSourceUnpackedness NoSourceStrictness
+                    , VarT (mkName "b")
+                    )
+                  ]
                   ( mkName "T" )
                   [ ConT (mkName "Int") ] ]
            [] ])