+++ testsuite/tests/th/T10828b.hs
@@ -10,16 +10,30 @@ $( return
    [ DataD [] (mkName "T")
            [ PlainTV (mkName "a") ]
            (Just StarT)
-           [ NormalC (mkName "MkT")
-                   [ (NotStrict, VarT (mkName "a"))
-                   , (NotStrict, VarT (mkName "a"))]
+           [ NormalC
+               (mkName "MkT")
+               [ ( Bang NoSourceUnpackedness NoSourceStrictness
+                 , VarT (mkName "a")
+                 )
+               , ( Bang NoSourceUnpackedness NoSourceStrictness
+                 , VarT (mkName "a")
+                 )
+               ]
            , ForallC [PlainTV (mkName "a")]
                      [AppT (AppT EqualityT (VarT $ mkName "a"  ) )
                                            (ConT $ mkName "Int") ] $
-             RecGadtC [(mkName "MkC")]
-                  [ (mkName "foo", NotStrict, VarT (mkName "a"))
-                  , (mkName "bar", NotStrict, VarT (mkName "b"))]
-                  ( mkName "T" )
-                  [ ConT (mkName "Int") ]
+             RecGadtC
+                 [ (mkName "MkC")]
+                 [ ( mkName "foo"
+                   , Bang NoSourceUnpackedness NoSourceStrictness
+                   , VarT (mkName "a")
+                   )
+                 , ( mkName "bar"
+                   , Bang NoSourceUnpackedness NoSourceStrictness
+                   , VarT (mkName "b")
+                   )
+                 ]
+                 ( mkName "T" )
+                 [ ConT (mkName "Int") ]
            ]
            [] ])