+++ testsuite/tests/th/T5290.hs
@@ -5,4 +5,5 @@ module T5290 where
 import Language.Haskell.TH
 
 $( let n = mkName "T"
-   in return [DataD [] n [] Nothing [NormalC n [(Unpacked,ConT ''Int)]] []] )
+   in return [DataD [] n [] Nothing
+                [NormalC n [(Bang SourceUnpack SourceStrict,ConT ''Int)]] []] )